honksquad: fix: predict cybernetic eye flash immunity client-side
Move the FlashAttemptEvent handler that grants flash immunity to bodies
with advanced cybernetic eyes from server-only CyberneticOrganEffectsSystem
into a new shared SharedCyberneticEyesSystem. The wearer's client now
sees the cancellation immediately on its prediction tick, instead of
briefly rendering the flash overlay before the server's cancel state
snaps it back to clear.

No behavior change for non-wearers or for the server-side flash logic;
this is purely about predicting the cancel.

The existing CyberneticEyesEffectsTest fires FlashAttemptEvent on the
body and checks Cancelled, which still passes against the new shared
handler.
